Initial move of analysis module to transit_fitting with replacement of analysis imports.
