+++ setup.cfg
@@ -50,7 +50,7 @@ stellarphot.io.tests = data/*
 stellarphot.tests =
     coveragerc
     data/*
-stellarphot.analysis = data/*
+stellarphot.transit_fitting = data/*
 stellarphot.notebooks.photometry = *.ipynb
 
 [tool:pytest]
