--- stellarphot/analysis/__init__.py
@@ -1,2 +1,0 @@
-from .exotic import *
-from .transit_fitting import *
