+++ stellarphot/transit_fitting/__init__.py
@@ -1,0 +1,2 @@
+from .exotic import *
+from .model_fit import *
