+++ stellarphot/transit_fitting/data/exotic-to-mod-candidate.json
