+++ stellarphot/transit_fitting/data/exotic-to-mod-known.json
