+++ stellarphot/transit_fitting/data/tic-template-for-exotic-candidate.json
