+++ stellarphot/transit_fitting/data/tic-template-for-exotic-known.json
